support for systems with no std::locale named "C"
constructor discussion for std::locale here: http://en.cppreference.com/w/cpp/locale/locale/locale discusses the case where a system may not have a locale of the given `std_name`, and the ctor will throw an exception if asked to construct a locale with an std_name that does not match any available.  Google's Native Client environment is a system with no locale named "C".  This change simply protects from that case and ignores the attempt to reset the std::stringstream